requirements.txt
arrow==0.17.0
Django==2.2.19
django-cors-headers==3.5.0
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
